move all plugins to Lolcommits::Plugin namespace
