--- lib/core_ext/class.rb
@@ -1,7 +1,0 @@
-class Class
-  def subclasses
-    result = []
-    ObjectSpace.each_object(Class) { |klass| result << klass if klass < self }
-    result
-  end
-end
