+++ lib/lolcommits.rb
@@ -1,6 +1,5 @@
 $LOAD_PATH.unshift File.expand_path('.')
 
-require 'core_ext/class'
 require 'mini_magick'
 require 'core_ext/mini_magick/utilities'
 require 'fileutils'
@@ -16,7 +15,6 @@ require 'lolcommits/configuration'
 require 'lolcommits/capturer'
 require 'lolcommits/vcs_info'
 require 'lolcommits/installation'
-require 'lolcommits/plugin'
 require 'lolcommits/platform'
 
 # after lolcommits/platform, so that we can do platform-conditional override
@@ -28,19 +26,20 @@ require 'lolcommits/backends/installation_mercurial'
 require 'lolcommits/backends/git_info'
 require 'lolcommits/backends/mercurial_info'
 
-require 'lolcommits/plugins/loltext'
-require 'lolcommits/plugins/dot_com'
-require 'lolcommits/plugins/tranzlate'
-require 'lolcommits/plugins/lol_twitter'
-require 'lolcommits/plugins/uploldz'
-require 'lolcommits/plugins/term_output'
-require 'lolcommits/plugins/lolsrv'
-require 'lolcommits/plugins/lol_yammer'
-require 'lolcommits/plugins/lol_protonet'
-require 'lolcommits/plugins/lol_tumblr'
-require 'lolcommits/plugins/lol_slack'
-require 'lolcommits/plugins/lol_flowdock'
-require 'lolcommits/plugins/lol_hipchat'
+require 'lolcommits/plugin/base'
+require 'lolcommits/plugin/loltext'
+require 'lolcommits/plugin/dot_com'
+require 'lolcommits/plugin/tranzlate'
+require 'lolcommits/plugin/lol_twitter'
+require 'lolcommits/plugin/uploldz'
+require 'lolcommits/plugin/term_output'
+require 'lolcommits/plugin/lolsrv'
+require 'lolcommits/plugin/lol_yammer'
+require 'lolcommits/plugin/lol_protonet'
+require 'lolcommits/plugin/lol_tumblr'
+require 'lolcommits/plugin/lol_slack'
+require 'lolcommits/plugin/lol_flowdock'
+require 'lolcommits/plugin/lol_hipchat'
 
 # require runner after all the plugins have been required
 require 'lolcommits/runner'
